Aim1/DeepMHCI/requirements.txt
torch==1.7.1
numpy==1.19.2
scipy==1.6.1
scikit-learn==0.24.1
click==7.1.2
ruamel.yaml==0.16.12
tqdm==4.56.0
logzero==1.6.3
